% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyTheilSen.R
\name{apply.Theil.Sen}
\alias{apply.Theil.Sen}
\title{Gebruik functie voor het vaststellen van de robuuste
Theil Sen helling.}
\usage{
apply.Theil.Sen(y, Series)
}
\description{
Gebruik functie voor het vaststellen van de robuuste
Theil Sen helling.
}
