% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{basemap}
\alias{basemap}
\title{Maakt ggplot2 object met LMG basemap}
\usage{
basemap()
}
\value{
een ggplot2 object
}
\description{
Deze functie maakt een ggplot2 object met de basemap met
de grondwaterlichamen in Nederland. Ruimtelijke data kan
worden toegevoegd aan deze plot met de ggplot2 syntax.
}
\examples{

p <- basemap()
p <- p + geom_sf(data=pts) 

}
