% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhfdr.R
\name{bhfdr}
\alias{bhfdr}
\title{Berekening False Discovery Rate (FDR)}
\usage{
bhfdr(p, alpha = 0.05)
}
\arguments{
\item{p}{vector van p-waardes}

\item{alpha}{alpha waarde voor FDR procedure, standaard is 0.05}
}
\value{
lijst met de resultaten van de FDR procedure
 \itemize{
   \item rejects Hoeveelheid afgewezen H0's
   \item threshold Max p-waarde van de procedure
   \item gamma.hat p-waarde van de minst siginificante p-waarde 
   gebaseerd op dichtheid
   \item df data frame met alle data, gebruikt voor de plot functies
}
}
\description{
Deze functie berekent de Benjamini-Hochberg False Discovery
Rate (FDR).
}
