% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgwl.R
\name{getgwl}
\alias{getgwl}
\title{Help functie voor een lijst van de grondwaterlichamen}
\usage{
getgwl()
}
\value{
een lijst van grondwaterlichaamnamen uit de LMG dataset
}
\description{
Help functie voor een lijst van de grondwaterlichamen
}
