% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetverdeling2.R
\name{meetverdeling2}
\alias{meetverdeling2}
\title{Toets Meetverdeling2}
\usage{
meetverdeling2(d, param, diep)
}
\arguments{
\item{d}{tijdreeks van waarnemingen voor 1 parameter en 1 diepte}
}
\value{
een data.frame met het aantal metingen per KRW planperiode
om te toetsen aan de verdeling van het aantal metingen, de volgende
kolommen worden gegeven:
 \itemize{
   \item aantal.metingen aantal metingen per planperiode
   \item totaal totaal aantal metingen per put
}
}
\description{
Deze functie bepaalt de meetverdeling door de jaren voor de
verschillende KRW planperiodes op putfilterniveau
}
