% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monsterFreq.R
\name{monsterFreq}
\alias{monsterFreq}
\title{Maakt een overzicht van de bemonsteringsfrequentie}
\usage{
monsterFreq(x, param)
}
\arguments{
\item{x}{grondwaterdataset.}
}
\value{
data.frame met bemonsteringsfrequenties
}
\description{
Per grondwaterlichaam en filterdiepte wordt een overzicht gemaakt
van het aantal filters per aantal bemonsteringen. Dit gebeurt op 
het niveau van één vaak bemonsterde stof, bijvoorbeeld no3.
}
