% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all.R
\name{plot_all}
\alias{plot_all}
\title{Maakt tijdreeksgrafieken van de gegevens}
\usage{
plot_all(
  i,
  x,
  trim = FALSE,
  trimfactor = 1.5,
  dw.plot = TRUE,
  rpDL = TRUE,
  replacefactor = 0.5
)
}
\arguments{
\item{i}{put/filter naam waarvoor de statistiek berekend wordt}

\item{x}{data.frame van lmgsubset}

\item{trim}{als TRUE, dan worden uitbijters verwijderd (trimmed)}

\item{dw.plot}{geeft aan om de relevante drempelwaarde in de grafiek wordt geplot}

\item{rpDL}{of de rapportagegrenzen aangepast moeten worden, standaard op TRUE.}

\item{trimFactor}{factor van \code{rmoutlier}}
}
\value{
een plot object
}
\description{
De data van de \code{lmgsubset} wordt omgezet in een tijdreeks,
uitbijters worden verwijderd met \code{rmoutlier}. Elke reeks wordt geplot,
ongeacht het aantal metingen.
}
