% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNegative.R
\name{replaceNegative}
\alias{replaceNegative}
\title{Vervang waardes onder de detectielimiet.}
\usage{
replaceNegative(x, replaceval = NA)
}
\arguments{
\item{x}{vector met waardes}

\item{replaceval}{vermenigvuldigingsfactor om negatieve waardes
te vervangen}
}
\value{
een vector met waardes waar de detectielimiet is vervangen
}
\description{
Help functie voor \code{lmgsubset} om waardes onder de 
detectielimiet te vervangen.
}
\details{
Deze functie vervangt waardes onder de detectielimiet. Deze 
waardes zijn gecodeerd als negatieve concentraties in de LMG
dataset. Om de detectielimet te vervangen voor de helft van de
oorspronkelijke waarde moet de vermenigvuldigingsfactor 0,5 zijn.
}
\examples{
data(lmg)
x<-replaceNegative(lmg$no3,-0.5)

}
