% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmoutlier.R
\name{rmoutlier}
\alias{rmoutlier}
\title{Tukey's methode om uitbijters te verwijderen.}
\usage{
rmoutlier(x, factor = 1.5, na.rm = FALSE)
}
\arguments{
\item{x}{vector van waardes}

\item{factor}{factor om de interkwartielafstand (IQR) uit te breiden}

\item{na.rm}{verwijder NA waardes}
}
\value{
vector met uitbijters vervangen door NA
}
\description{
Gebruik Tukey's methode om uitbijters te markeren als NA. Uitbijters zijn
waardes welke groter of kleiner zijn dan 1,5 (factor) keer de 
interkwartielafstand (IQR) plus of min kwartiel 2 of 4. NA waardes worden
niet automatisch verwijderd en resulteren in een error.
}
