% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toetsNormoverschrijding.R
\name{toetsNormoverschrijding}
\alias{toetsNormoverschrijding}
\title{Toets Normoverschrijding}
\usage{
toetsNormoverschrijding(d, gwl, toetsnorm = 0.75)
}
\arguments{
\item{d}{tijdreeks van waarnemingen voor 1 parameter en 1 diepte}

\item{gwl}{het grondwaterlichaam}

\item{toetsnorm}{de relatieve norm waar tegen getoetst wordt.

De normen zijn afhankelijk van het grondwaterlichaam waarvoor
getoetst wordt. De drempelwaarden per grondwaterlichaam zijn
gegeven in `data(drempelwaarden)`.

De toetstnorm is de relatieve norm. Deze is standaard 0.75 om te
toetsen aan 75% van de norm. Als getoetst moet worden t.o.v. de 
actuele drempelwaarde, dan moet de toetsnorm op 1 gezet worden.}
}
\value{
een data.frame met de toets statistieken, de volgende
kolommen worden gegeven:
 \itemize{
   \item n totaal aantal gegeven meetwaarden boven de toetsnorm
   \item percentage percentage boven de toetsnorm
   \item gem.waarde gemiddelden van waarden boven de toetsnorm
   \item gem.ratio gemiddelde ratio van waarde/norm
}
}
\description{
Deze functie berekent het percentage van de  waarnemingen met
normoverschrijdingen. Het berekent ook de gemiddelde concentratie
van de putten met normoverschrijdingen en de gemiddelde ratio van
concentratie (waarde) gedeeld door de norm.
}
