% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toetsStatistiek.R
\name{toetsStatistiek}
\alias{toetsStatistiek}
\title{Toets Statistiek}
\usage{
toetsStatistiek(d)
}
\arguments{
\item{d}{tijdreeks van waarnemingen voor 1 parameter en 1 diepte}
}
\value{
een data.frame met de toets statistieken, de volgende
kolommen worden gegeven:
 \itemize{
   \item ntot totaal aantal gegeven meetwaarden zonder NA's
   \item P50 mediaan van meetwaarden
   \item P95 95ste percentiel van meetwaarden
   \item perc.RG percentage meetwaarden onder de rapportagegrens (RG)
   \item RG.min laagste rapportagegrens
   \item RG.max hoogste rapportagegrens

}
}
\description{
Deze functie geeft een overzicht van de statistieken in een 
grondwaterdataset. Het 50ste en 95ste percentiel worden
berekend, waarbij zowel metingen <RG als erboven zijn meegenomen.
Daarnaast wordt het percentage metingen onder de rapportagegrens 
bepaald, met de laagste en hoogste rapportagegrens binnen de dataset.
}
