% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendomkeringAnalyse.R
\name{trendomkeringAnalyse}
\alias{trendomkeringAnalyse}
\title{Bereken gegevens trendomkering}
\usage{
trendomkeringAnalyse(d, sig = 0.05)
}
\arguments{
\item{sig}{alpha waarde (significantie) van trendomkeringen}

\item{x}{data.frame met grondwaterdata}
}
\value{
tabel met trendstatistieken
 \itemize{
   \item aantal trendomkeringen
   \item percentage trendomkeringen
   \item gemiddelde hellingen van trendtrendomkeringen
   \item gemiddeld jaar en standaardafwijking van keerpunt
   \item voor benoemde als niet benoemde trends
}
}
\description{
Deze functie berekent de statistieken voor de trendomkering. Het
bepaald of trendomkeringen aanwezig zijn en het jaar waarin de
omkering plaatsvindt.
}
\details{
De trends worden gecorrigeerd voor vals positieven.
}
