% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendReversal.R
\name{trendReversal}
\alias{trendReversal}
\title{Trend reversal test}
\usage{
trendReversal(
  i,
  x,
  trim = FALSE,
  rpDL = TRUE,
  trimfactor = 1.5,
  make.plot = FALSE,
  dw.plot = TRUE
)
}
\arguments{
\item{i}{put/filter naam waar statistiek voor berekend wordt}

\item{x}{data.frame van lgmsubset}

\item{trim}{als TRUE, dan worden uitbijters verwijderd (trimmed)}

\item{make.plot}{als TRUE dan wordt een plot object aangemaakt}

\item{trimFactor}{factor van \code{rmoutlier}}
}
\value{
een trendReversal object met daarin de resultaten van de
trend reversal test en, als make.plot is TRUE, een ggplot object
met een grafische weergave van de trendomkering

Een trendomkering ontstaat als de concentratie (van een stof in een
put) vanaf een bepaalde moment/jaar - het keerpunt genoemd -
toeneemt (of afneemt, of constant is) terwijl de concentratie
daarvoor afnam (of toenam, of constant was). Het detecteren van een
trendomkering is meer complex dan het detecteren van een trend
alleen, aangezien dat er meerdere patronen zijn waarop een reeks
geen duidelijk toenemende or afnemende trend aantoont en waardoor
een trendomkering onterecht gesignaleerd kan worden. 

Om zulke
patronen uit te sluiten wordt eerst gecontroleerd of de reeks
compatibel is met twee achtervolgende trends in verschillende
richtingen. Voor elke keuze van het potentiele keerpunt wordt de
reeks in twee stukken verdeeld, op elk stuk wordt een rechte lijn
aangepast door de robuuste methode van Theil-Sen, en een algemeen
mate van discrepantie tussen de lijnen en de werkelijke
concentraties berekent (namelijk de som van de kwadraten van de
‘residuen’, ofwel de verschillen tussen de aangepaste lijnen en de
concentraties). Het keerpunt met de kleinste discrepantie wordt
gekozen, met de bijbehorende lijnen. Alleen als de hellingen van de
lijnen van teken verschillen (een is <0 en de andere >0) gaat men
verder om een p-waarde te berekenen (welke p-waarde de aanwijzing
geeft voor het ontstaan van een trendomkering eerder dan een
simpele trend); als dit niet het geval is dan wordt zonder meer de
p-waarde gedefinieerd als 1 (wat geen aanwijzing voor trendomkering
geeft). 
Voor het berekenen van de p-waarde wordt een regressie
model met een kwadratische term aan de data gepast (m.a.w.: men
past een parabool i.p.v. een rechte lijn op de reeks
concentraties). Daarna wordt er verifieerd of het hoogste/laagste
punt van de aangepaste kromme binnen het tijdsvenster van de reeks
valt (zoals het hoort bij een trendomkering). In het geval dat het
niet zo is wordt de p-waarde gedefinieerd als 1. Anders wordt er
getoetst of het kwadratische model de data niet ‘aanzienlijk beter
beschrijft’ dan het simpeler recht lijnig model.
}
\description{
De trend reversal test is een methode om te toetsen of er
trendomkering in de data aanwezig is
}
