% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw.R
\name{dw}
\alias{dw}
\title{Selecteer drempelwaardes (DW)}
\usage{
dw(param, gwl, data = drempelwaarden)
}
\arguments{
\item{param}{parameter naam (als string)}

\item{gwl}{grondwaterlichaam (groundwaterbody)}

\item{data}{dataset met drempelwaardes. default is drempelwaardes dataset die
bij het pakket zit}
}
\value{
drempelwaarde per stof per grondwaterlichaam
}
\description{
Selecteer drempelwaardes (limit values)
}
\examples{
get list of groundwater bodies
getgwl()
dw("no3","NLGW0012")

}
