% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktrends.R
\name{mktrends}
\alias{mktrends}
\title{Berekent Mann-Kendall statistiek voor enkel LMG filter}
\usage{
mktrends(
  i,
  x,
  minwaarnemingen = 10,
  trim = FALSE,
  trimfactor = 1.5,
  dw.plot = TRUE,
  psig = 0.05,
  alter = "two.sided",
  rpDL = TRUE,
  make.plot = FALSE,
  replacefactor = 0.5
)
}
\arguments{
\item{i}{put/filter naam waarvoor de statistiek berekend wordt}

\item{x}{data.frame van lmgsubset}

\item{minwaarnemingen}{minimum aantal waarnemingen waar een reeks aan moet 
voldoen, standaard op 10.}

\item{trim}{als TRUE, dan worden uitbijters verwijderd (trimmed)}

\item{dw.plot}{geeft aan om de relevante drempelwaarde in de grafiek wordt geplot}

\item{psig}{significanctie niveau}

\item{alter}{alternatieve hypothese van \code{kendatTrendTest}}

\item{rpDL}{of de rapportagegrenzen aangepast moeten worden, standaard op TRUE.}

\item{make.plot}{geeft een plot output in plaats van een data.frame}

\item{replacefactor}{vermenigvuldigingsfactor om rapportagegrens waardes
te vervangen}

\item{trimFactor}{factor van \code{rmoutlier}}
}
\value{
data.frame met statistiek of een plot object (als #' make.plot=TRUE)
}
\description{
Voor een enkele put/filter in de data van \code{lmgsubset}
wordt de Mann-Kendall statistiek berekend voor een parameter.
De data van de \code{lmgsubset} wordt omgezet in een tijdreeks,
uitbijters worden verwijderd met \code{rmoutlier}
}
