% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceDL.R
\name{replaceDL}
\alias{replaceDL}
\title{Vervang waardes onder de detectielimiet.}
\usage{
replaceDL(d, replaceval = 0.5)
}
\arguments{
\item{d}{vector met waardes}

\item{replaceval}{vermenigvuldigingsfactor om rapportagegrens waardes
te vervangen}
}
\value{
een vector met waardes waar de detectielimiet is vervangen
}
\description{
Help functie voor \code{lmgsubset} om waardes onder de 
detectielimiet te vervangen.
}
\details{
Deze functie vervangt waardes onder de detectielimiet. Deze 
waardes zijn gecodeerd door 1 in de detectielimiet kolom.
Om de detectielimet te vervangen voor de helft van de
oorspronkelijke waarde moet de vermenigvuldigingsfactor 0,5 zijn.
}
\examples{
data(lmg_reeks)
d <- replaceDL(lmg_reeks, 0.5)

}
