% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendAnalyse.R
\name{trendAnalyse}
\alias{trendAnalyse}
\title{Bereken gegevens trendanalyse}
\usage{
trendAnalyse(d, n.reeks)
}
\arguments{
\item{d}{data.frame met grondwaterdata}

\item{n.reeks}{totaal aantal reeksen van selectie}
}
\value{
tabel met trendstatistieken
 \itemize{
   \item aantal op- en neergaande trends
   \item percentage op- en neergaande trends t.o.v. totaal aantal reeksen en 
         t.o.v. aantal goedgekeurde reeksen.
   \item gemiddelde helling van trends
   \item voor benoemde als niet benoemde trends
}
}
\description{
Deze functie berekent de statistieken voor de trendanalyse. Het
bepaalt of trends aanwezig zijn en de richting van trends.
De trends worden gecorrigeerd voor vals positieven.
}
